% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivSurvey.R
\name{equivSurvey}
\alias{equivSurvey}
\title{equivSurvey function}
\usage{
equivSurvey(delta_l, delta_u, design12, indicator, subset)
}
\arguments{
\item{delta_l}{The lower bound of the equivalence test, chosen by the user}

\item{delta_u}{The upper bound of the equivalence test, chosen by the user}

\item{design12}{The survey design, obtained from svydesign function}

\item{indicator}{The outcome of interest as a string}

\item{subset}{Column name for survey indicator}
}
\value{
Table with equivalence test results
}
\description{
Test equivalence of two non-overlaping survey samples
}
\details{
This function will test the equivalence of two non-overlapping survey samples
}
\examples{
# equivSurvey(-0.05, 0.05, design12, 'fp19b', 'inMethod1OR2')

}
